requirements.txt
uvicorn
requests
deepgram-sdk==3.11.0
python-dotenv==1.0.0
openai==1.74.0
broadcaster==0.2.0
fastapi==0.95.0
jinja2==3.1.2
motor>=3.7.1